requirements.txt
cx_Oracle==8.3.0
pandas==2.1.4
reportlab==4.0.8
selenium==4.16.0
webdriver-manager==4.0.1
python-dotenv==1.0.0
